% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{archive_data}
\alias{archive_data}
\title{Archive data with date and version tag.}
\usage{
archive_data(path, files, save_path)
}
\arguments{
\item{path, }{a string: ending with "/". The source directory.}

\item{files, }{a list of characters: filenames or "all"}

\item{save_path, }{a string: ending with "/". The target directory.}
}
\description{
Archive data with date and version tag.
}
