% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{ccall}
\alias{ccall}
\title{Function to remove all variables and close all graphs/ plots
including Garbage Collection gc()}
\usage{
ccall()
}
\description{
Function to remove all variables and close all graphs/ plots
including Garbage Collection gc()
}

