% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{clc}
\alias{clc}
\title{Function to remove all variables and close all graphs/ plots
including Garbage Collection gc()}
\usage{
clc()
}
\description{
Function to remove all variables and close all graphs/ plots
including Garbage Collection gc()
}
