% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{copy_rename_file}
\alias{copy_rename_file}
\title{Copy and rename one file. Used in archive_data}
\usage{
copy_rename_file(path, file, save_path, prefix = "")
}
\arguments{
\item{path, }{a string: ending with "/"}

\item{prefix, }{a character added to the target filename}

\item{files, }{a list of characters: filenames or "all"}
}
\description{
Copy and rename one file. Used in archive_data
}
