% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{decimalplaces}
\alias{decimalplaces}
\title{decimalplaces returns the number of decimal places of x}
\usage{
decimalplaces(x)
}
\arguments{
\item{x}{A numeric}
}
\value{
result An integer, the number of decimalplaces of x
}
\description{
decimalplaces returns the number of decimal places of x
}
