% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{dict}
\alias{dict}
\title{Define a German <-> English dictionary for weekday and month names SLC: Remove???}
\usage{
dict()
}
\value{
A dictionary as list (subset possible)
}
\description{
Define a German <-> English dictionary for weekday and month names SLC: Remove???
}
