% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{dir.remove}
\alias{dir.remove}
\title{Remove directory}
\usage{
dir.remove(path, recursive, force)
}
\arguments{
\item{path}{A character, the path}

\item{recursive}{A boolean, set to TRUE to remove directories.
Refer to help for unlink.}

\item{force}{A boolean, refer to help for unlink.}
}
\value{
TRUE value which can be assigned, but which is not print. ?invisible.
}
\description{
Wrapper for unlink to automatically remove directories. For details see unlink.
}
