% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{e2g}
\alias{e2g}
\alias{e2g}
\title{e2g english to german using weekdays_abbr as reference. SLC: Remove???}
\usage{
e2g(eng_expr)

e2g(eng_expr)
}
\arguments{
\item{eng_expr}{A character (in english)}

\item{weekday_eng}{A character}
}
\value{
res

A character, the translation
}
\description{
e2g english to german using weekdays_abbr as reference. SLC: Remove???

English => German translator based on dict() SLC: Remove???
}
