% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{googlestyle}
\alias{googlestyle}
\title{Function without functionality to show R style guideline from google}
\usage{
googlestyle()
}
\description{
Google R styleguide, see "https://google.github.io/styleguide/Rguide.xml#generallayout":
}
\details{
Files end with .R: meaningful_file.R
Variables: meaningful.variable
Functions: MeaningfulFunction

Indentation
When indenting your code, use two spaces.  Never use tabs or mix tabs and spaces.
Spacing
Place spaces around all binary operators (=, +, -, <-, etc.).
Do not place a space before a comma, but always place one after a comma.
Place a space before left parenthesis, except in a function call.

General Layout and Ordering
1. Copyright statement comment
2. Author comment
3. File description comment, including purpose of program, inputs, and outputs
4. source() and library() statements
5. Function definitions
6. Executed statements, if applicable (e.g., print, plot)

Unit tests should go in a separate file named originalfilename_test.R.

Install packages with install.packages("package")
Include non-standard libraries using libraray(package)
}

