% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{\%<-\%}
\alias{\%<-\%}
\title{%<-% Matlabs '[...] =' operator. call as c(par1, par2, ...) %<-% somefunction(...)
instead of calling res <- somefunction(...), par1 <- res$par1, ... rm(res)
Should not be used in simulations}
\usage{
lhs \%<-\% rhs
}
\arguments{
\item{rhs}{}
}
\value{
result
}
\description{
%<-% Matlabs '[...] =' operator. call as c(par1, par2, ...) %<-% somefunction(...)
instead of calling res <- somefunction(...), par1 <- res$par1, ... rm(res)
Should not be used in simulations
}
