% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{ims_match}
\alias{ims_match}
\title{SQC's strange but working match function}
\usage{
ims_match(pattern, y, encoding_match_type = "base_r")
}
\arguments{
\item{pattern}{the pattern to be matched}

\item{y}{the vector being searched}
}
\value{
array of indices with exact case sensitive matching
}
\description{
Search for a string (case sensitive, not a pattern!) in a vector y. Full match
is done through comparison of string's length. Used in replace_by_lookuptable()
In theory, this function should be identical to match(pattern, y).
Sometimes, match has problems with encodings. Not reproducible.
}
\details{
Comment: Method has been taken from zh911::my_match and renamed as ims_match.
}
\examples{
y <- c("Test", "a", "b", "t", "T", "Te", "Test", "hurz")
ims_match("T", y) # 5
y[ims_match("T", y)] # "T"
grep("T", y) # 1 5 6 7
grepl("T", y)  #  TRUE FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE => returns too much
}
