% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{install_simiverse}
\alias{install_simiverse}
\title{install_simiverse: Installs all the packages for the sim911-universe.
Uses devtools::install_github, accesses always master branches.
No examples and tests provided since they will change your library..}
\usage{

  install_simiverse(auth_token = stop("Provide GitHub AuthToken to access private Repos"))
}
\arguments{
\item{auth_token}{The token to access your privat Repos on GitHub}
}
\value{
df_package_status A data.frame with package names and installation status
}
\description{
install_simiverse: Installs all the packages for the sim911-universe.
Uses devtools::install_github, accesses always master branches.
No examples and tests provided since they will change your library..
}
