% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{int_dict}
\alias{int_dict}
\title{Define a Integer-alias <-> English dictionary for weekday and month names ???}
\usage{
int_dict()
}
\value{
A dictionary as list (subset possible)
}
\description{
Define a Integer-alias <-> English dictionary for weekday and month names ???
}
