% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{is.installed}
\alias{is.installed}
\title{Check if package is installed}
\usage{
is.installed(mypkg)
}
\arguments{
\item{mypkg}{A character, the package name}
}
\value{
A boolean, TRUE or FALSE
}
\description{
Check if package is installed
}
\examples{
is.installed("imsbasics") #  [1] TRUE
}
