% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{load_rdata}
\alias{load_rdata}
\title{Load data from specific folder to a variable in current environment.
Call as data <- imsbasics::load_rdata(filename, path)}
\usage{
load_rdata(filename, path = stop("Filename and path arguments are required"))
}
\arguments{
\item{filename, }{can be without ".RData"}

\item{path}{in "../.." style ending with "/".}
}
\value{
Obects(s) in filename
}
\description{
Load data from specific folder to a variable in current environment.
Call as data <- imsbasics::load_rdata(filename, path)
}
