% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{midpoints}
\alias{midpoints}
\title{midpoints calculates the midpoints of a factor level (x,y] created from ..... ?}
\usage{
midpoints(x, dp = 2)
}
\arguments{
\item{dp}{}
}
\value{
mids
}
\description{
midpoints calculates the midpoints of a factor level (x,y] created from ..... ?
}
