% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{percent_deviation}
\alias{percent_deviation}
\title{Return deviation of a number from a reference number x_ref in percent.}
\usage{
percent_deviation(x, x_ref, digits = 1)
}
\arguments{
\item{x, }{a numeric}

\item{x_ref, }{a numeric}
}
\value{
absolute value of deviation in percent
}
\description{
Return deviation of a number from a reference number x_ref in percent.
}
