% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{plot_runtime}
\alias{plot_runtime}
\title{Plot runtime graph on a double log scale}
\usage{
plot_runtime(n, dt, dt_unit, title, display)
}
\arguments{
\item{n}{iterations}

\item{dt}{runtime}

\item{dt_unit}{A character, the unit of runtime}

\item{title}{A character, the title of the graph}

\item{display}{A bool, display fit if T}
}
\description{
Plot runtime graph on a double log scale
}
\examples{
n <- 10^c(0:5)
dt <- 300*n+50 # in us
plot_runtime(n, dt, "us", "My function", T)
}
