% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{pre_commit}
\alias{pre_commit}
\title{pre_commit: uses devtools to document package, run all examples and
run all testthat tests and if use_r_cmd_check = T even run devtools::check.
Use this before every commit.}
\usage{
pre_commit(use_r_cmd_check = T)
}
\arguments{
\item{use_r_cmd_check}{A boolean if TRUE, devtools::check() will be used}
}
\description{
R CMD check() can be set inactive (use_r_cmd_check = F), because it may
throw errors even if devtools::run_examples() and devtools::test()
run without errors. Not fully understood by SCN.
}
