% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{r_options}
\alias{r_options}
\title{set_custom_rstudio sets custom parameters}
\usage{
r_options(error = NULL, warn = 0, strings_as_factors = F, english = T)
}
\arguments{
\item{warn}{= 0}
}
\description{
set_custom_rstudio sets custom parameters
}
