% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{shift_array}
\alias{shift_array}
\title{shift array by n positions to the right or left}
\usage{
shift_array(x, n, default = NA)
}
\arguments{
\item{x}{An array}

\item{n}{An integer, the number of shifts}

\item{default}{NA or any other type. which should be placed where i < n.}
}
\value{
An array
}
\description{
shift array by n positions to the right or left
}
\examples{
shift_array(c(1, 2, 3, 4), 2)
shift_array(c(1, 2, 3, 4), -2, "hurz")
}
