% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{weekdays_abbr}
\alias{weekdays_abbr}
\title{weekdays_abbr dictionary according to lubridate}
\usage{
weekdays_abbr()
}
\value{
german and english weekdays
}
\description{
weekdays_abbr dictionary according to lubridate
}
