% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{zero_n}
\alias{zero_n}
\title{zero_n creates string of form 000x from integer x.}
\usage{
zero_n(number, position = 3)
}
\arguments{
\item{number}{A numeric, the number}

\item{position}{An integer, the number of leading zeros.}
}
\value{
A character, a standardized string of number
}
\description{
zero_n creates string of form 000x from integer x.
}
\examples{
imsbasics::zero_n(1.234, 1) # [1] "1.234"
imsbasics::zero_n(1.234, 2) # [1] "01.234"
imsbasics::zero_n(1.234, 3) # [1] "001.234"
}
