% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{g2e}
\alias{g2e}
\title{g2e german to english using weekdays_abbr as reference. SLC: Remove???}
\usage{
g2e(ger_expr)

g2e(ger_expr)
}
\arguments{
\item{ger_expr}{A character (in german)}

\item{weekday_ger}{A character}
}
\value{
res

A character, the translation
}
\description{
g2e german to english using weekdays_abbr as reference. SLC: Remove???

German => English translator based on dict() SLC: Remove???
}
