% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{midpoints}
\alias{midpoints}
\title{Calculates the midpoints of a factor level (x,y]}
\usage{
midpoints(x, digits = 2)
}
\arguments{
\item{x}{A factor}

\item{digits}{An integer, the number of digits}
}
\value{
mids A numeric, the array of midpoints
}
\description{
Calculates the midpoints of a factor level (x,y]
}
\examples{
n <- 3; k <- 5
d <- gl(n, k, length = n*k, labels = seq_len(n), ordered = FALSE)
x <- cut(as.numeric(d), 4)
imsbasics::midpoints(x, 4)
}
