% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{rectangle}
\alias{rectangle}
\title{rectangle step to zero in (t0, t1). Otherwise 0.}
\usage{
rectangle(t, lower, upper, at_step = 0.5)
}
\arguments{
\item{t1}{}
}
\description{
rectangle step to zero in (t0, t1). Otherwise 0.
}
