% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{replace_by_lookuptable}
\alias{replace_by_lookuptable}
\title{replace entries using a lookup table}
\usage{
replace_by_lookuptable(df, col, lookup)
}
\arguments{
\item{df}{A data.frame}

\item{col}{A charcter vector, the name(s) of the columns with replacements}

\item{lookup}{A data.frame, the lookup table}
}
\value{
df with replaced entries
}
\description{
replace entries using a lookup table
}
\examples{
set.seed(1)
df <- data.frame(name = c(1:5),
                 x = sample(c("a", "b", "c"), 5, T),
                 y = sample(c("a", "b", "c"), 5, T), stringsAsFactors = F)
lookup <- data.frame(old = c("a", "b", "c"),
                     new = c("new_a", "new_b", "new_c"))
new_df <- replace_by_lookuptable(df, "x", lookup)
new_df <- replace_by_lookuptable(df, c("x", "y"), lookup)
}
