% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsbasics.R
\name{zero_n}
\alias{zero_n}
\title{zero_n creates string of form 000x from integer x.}
\usage{
zero_n(number, position = 3)
}
\arguments{
\item{position}{}
}
\value{
standard string
}
\description{
zero_n creates string of form 000x from integer x.
}
